+++ online-store/src/core/components/index.ts
@@ -1,11 +1,7 @@
-/* eslint-disable import/prefer-default-export */
-export { default as Header } from './header';
-export { default as Footer } from './footer';
-export { default as Main } from './main';
 export { default as Product } from './product';
 export { default as Settings } from './settings';
-export { default as RangeFilters } from './UI/rangeFilters';
-export { default as Buttons } from './UI/buttons';
-export { default as Search } from './UI/search';
-export { default as Sort } from './UI/sort';
-export { default as ValueFilters } from './UI/valueFilters';
+export { default as RangeFilters } from './ui/rangeFilters';
+export { default as Buttons } from './ui/buttons';
+export { default as Search } from './ui/search';
+export { default as Sort } from './ui/sort';
+export { default as ValueFilters } from './ui/valueFilters';
