+++ online-store/src/core/layouts/index.ts
@@ -1,0 +1,3 @@
+export { default as Header } from './header';
+export { default as Footer } from './footer';
+export { default as Main } from './main';
