+++ online-store/src/pages/index.ts
@@ -1,0 +1,4 @@
+export { default as App } from './app';
+export { default as MainPage } from './home';
+export { default as CartPage } from './cart';
+export { default as ErrorPage } from './error';
